requirements.txt
alembic==1.7.4
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.2
colorama==0.4.4
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
python-editor==1.0.4
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
sqlalchemy-cockroachdb==1.4.0
toml==0.10.2
Werkzeug==2.0.2

Manager~=2.0.3
Flask-Testing==0.8.1